Refined install script to allow for installation of offline mode, without cloud synced location logs.

+++ install.sh
@@ -19,20 +19,22 @@ echo $input>/home/pi/readykey/KeyID
 #step 4: run the rclone configuration script
 #currently must run be configured over vnc or with a display to log in to google drive via a browser
 # install raspberrypi-ui-mods and chromium-browser to raspian lite before running this script
-read -r -p "Do you want to configure rclone to sync location logs to Google Drive? 
-read -r -p "Do you have a desktop environment and browser running? [y/N] " response
-if [[ "$response" =~ ^([yY][eE][sS]|[yY])+$ ]]
-then
-    rclone config create cloud drive
+read -r -p "Do you want to configure rclone to sync location logs to Google Drive? [y/N] "
+if [[  "$answer" =~ ^([nN][oO][nN])+$ ]]
+then 
+	exit
+else
+	read -r -p "Do you have a desktop environment and browser running on this device? [y/N] " response
+	if [[ "$response" =~ ^([yY][eE][sS]|[yY])+$ ]]
+	then
+   		rclone config create cloud drive
 #step 5: configure crontab to run rclone every 15 mins. check the second answer
 #  https://stackoverflow.com/questions/8579330/appending-to-crontab-with-a-shell-script-on-ubuntu crontab -r
-    (crontab -l ; echo "0,15,30,45 * * * * /home/pi/readykey/logsync.sh")| crontab -
-    (crontab -l ; echo "0 1 * * * sudo find /home/pi/readykey/log/ -mtime +7 -type f -delete")| crontab -
-    (crontab -l ; echo "0 3 * * * sudo reboot")| crontab -
-else
-    exit
-fi      
-           
-
-
-
+    		(crontab -l ; echo "0,15,30,45 * * * * /home/pi/readykey/logsync.sh")| crontab -
+    		(crontab -l ; echo "0 1 * * * sudo find /home/pi/readykey/log/ -mtime +7 -type f -delete")| crontab -
+    		(crontab -l ; echo "0 3 * * * sudo reboot")| crontab -
+	else
+		read -r -p "Please install a desktop and browser, or configure rclone manaully."
+		exit
+	fi
+fi
